+++ babel.config.js
@@ -40,13 +40,6 @@ const jsdomTestPlugin = resolvePath(
 
 const aliases = {
   'effector/fixtures': resolveFromSources('fixtures'),
-  'effector-react/scope': {
-    esm: 'effector-react/scope.mjs',
-    compat: null,
-    build: null,
-    test: null,
-    default: null,
-  },
   'effector-react': {
     esm: 'effector-react/effector-react.mjs',
     compat: 'effector-react/compat',
@@ -55,13 +48,6 @@ const aliases = {
     default: resolveFromSources('../npm/effector-react'),
   },
   'effector-vue': resolveFromSources('vue'),
-  'effector-solid/scope': {
-    esm: 'effector-solid/scope.mjs',
-    compat: null,
-    build: null,
-    test: null,
-    default: null,
-  },
   'effector-solid': resolveFromSources('solid'),
   Builder: resolveFromSources('../tools/builder'),
   effector: {