+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 See also [separate changelogs for each library](https://changelog.effector.dev/)
 
+## effector 24.0.0
+
+- Delete deprecated option `reactSsr` in babel-plugin ([PR #1084](https://github.com/effector/effector/pull/1084))
+
+## effector-react 24.0.0
+
+- Delete deprecated `effector-react/scope` in favor of isomorphic hooks ([PR #1084](https://github.com/effector/effector/pull/1084))
+
+## effector-solid 24.0.0
+
+- Delete deprecated `effector-solid/scope` in favor of isomorphic hooks ([PR #1084](https://github.com/effector/effector/pull/1084))
+
+## effector-vue 24.0.0
+
+- Delete deprecated `effector-vue/ssr` in favor of isomorphic hooks ([PR #1084](https://github.com/effector/effector/pull/1084))
+- Delete deprecated `VueSSRPlugin` in favor of `EffectorScopePlugin` ([PR #1084](https://github.com/effector/effector/pull/1084))
+
 ## effector 23.2.2
 
 - Fix types for createEffect and attach in factories with generics ([issue #1069](https://github.com/effector/effector/issues/1069)). Cases like this now works correctly: